requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
Django==4.1
django-crispy-forms==1.14.0
django-filter==22.1
django-modeltranslation==0.18.4
django-rosetta==0.9.8
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Pillow==9.2.0
polib==1.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.2.0
